Add modules for functions
