+++ modules/game.js
@@ -1,34 +1,34 @@
-//Adding a name game
+//Adding a new game
 
 const addGame = async () => {
   try {
     const response = await fetch(`${API_URL}/games`, {
-      method: 'POST',
+      method: "POST",
       body: JSON.stringify({
-        "name": 'The best Game Ever',    
+        name: "the Best game in the world",
       }),
       headers: {
-        'Content-type': 'application/json; charset=UTF-8',
+        "Content-type": "application/json; charset=UTF-8",
       },
     });
-    
+
     const data = await response.json();
-    
+
     if (data && data.result) {
-      const gameId = data.result.match(/(\w+){10,}/ig);
+      const gameId = data.result.match(/(\w+){10,}/gi);
       return gameId[0];
     } else {
-      throw new Error('Error al obtener el ID del juego.');
+      throw new Error("Error al obtener el ID del juego.");
     }
   } catch (error) {
-    console.error('Error:', error);
+    console.error("Error:", error);
   }
-}
+};
 
 addGame()
-  .then(gameId => {
+  .then((gameId) => {
     console.log(gameId);
   })
-  .catch(error => {
-    console.error('Error:', error);
+  .catch((error) => {
+    console.error("Error:", error);
   });
